+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/dtos/LoginRequestDto.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.adapters.inbound.web.auth
+package app.xquare.xquareinfra.adapters.inbound.web.auth.dtos
 
 data class LoginRequestDto(
     val username: String,
