+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/dtos/RefreshTokenRequestDto.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.adapters.inbound.web.auth
+package app.xquare.xquareinfra.adapters.inbound.web.auth.dtos
 
 data class RefreshTokenRequestDto(
     val refreshToken: String,
