+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/dtos/TokenResponseDto.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.adapters.inbound.web.auth
+package app.xquare.xquareinfra.adapters.inbound.web.auth.dtos
 
 import app.xquare.xquareinfra.infrastructure.web.dto.SuccessResponseDto
