--- src/main/kotlin/app/xquare/xquareinfra/infrastructure/persistence/user/UserPersistenceRole.kt
@@ -1,6 +1,0 @@
-package app.xquare.xquareinfra.infrastructure.persistence.user
-
-enum class UserPersistenceRole {
-    ADMIN,
-    MEMBER,
-}
