+++ src/main/kotlin/app/xquare/xquareinfra/infrastructure/persistence/user/schema/UserPersistenceRole.kt
@@ -1,0 +1,6 @@
+package app.xquare.xquareinfra.infrastructure.persistence.user.schema
+
+enum class UserPersistenceRole {
+    ADMIN,
+    MEMBER,
+}
